moved API views and functions related to users page to separate file
