+++ poem/Poem/api/views_internal.py
@@ -13,7 +13,6 @@ from rest_framework_api_key import models as api_models
 from Poem.helpers.versioned_comments import new_comment
 from Poem.poem import models as poem_models
 from Poem.poem_super_admin.models import Probe, ExtRevision
-from Poem.users.models import CustUser
 from Poem.poem.saml2.config import tenant_from_request, saml_login_string, get_schemaname
 
 from reversion.models import Version, Revision
@@ -29,6 +28,7 @@ import json
 from Poem.api.internal_views.aggregationprofiles import *
 from Poem.api.internal_views.metricprofiles import *
 from Poem.api.internal_views.services import *
+from Poem.api.internal_views.users import *
 
 
 def get_groups_for_user(user):
@@ -253,88 +253,6 @@ class ListTokens(APIView):
             return Response(status=status.HTTP_400_BAD_REQUEST)
 
 
-class ListUsers(APIView):
-    authentication_classes = (SessionAuthentication,)
-
-    def get(self, request, username=None):
-        if username:
-            try:
-                user = CustUser.objects.get(username=username)
-                serializer = serializers.UsersSerializer(user)
-                return Response(serializer.data)
-
-            except CustUser.DoesNotExist:
-                raise NotFound(status=404,
-                            detail='User not found')
-
-        else:
-            users = CustUser.objects.all()
-            serializer = serializers.UsersSerializer(users, many=True)
-
-            return Response(serializer.data)
-
-    def put(self, request):
-        user = CustUser.objects.get(username=request.data['username'])
-        user.first_name = request.data['first_name']
-        user.last_name = request.data['last_name']
-        user.email = request.data['email']
-        user.is_superuser = request.data['is_superuser']
-        user.is_staff = request.data['is_staff']
-        user.is_active = request.data['is_active']
-        user.save()
-
-        return Response(status=status.HTTP_201_CREATED)
-
-    def post(self, request):
-        try:
-            CustUser.objects.create_user(
-                username=request.data['username'],
-                password=request.data['password'],
-                email=request.data['email'],
-                first_name=request.data['first_name'],
-                last_name=request.data['last_name'],
-                is_superuser=request.data['is_superuser'],
-                is_staff=request.data['is_staff'],
-                is_active=request.data['is_active']
-            )
-            user = CustUser.objects.get(username=request.data['username'])
-
-            userprofile = poem_models.UserProfile(
-                user=user,
-                displayname=request.data['displayname'],
-                subject=request.data['subject'],
-                egiid=request.data['egiid']
-            )
-            userprofile.save()
-
-            for group in request.data['groupsofaggregations']:
-                userprofile.groupsofaggregations.add(poem_models.GroupOfAggregations.objects.get(name=group))
-
-            for group in request.data['groupsofmetrics']:
-                userprofile.groupsofmetrics.add(poem_models.GroupOfMetrics.objects.get(name=group))
-
-            for group in request.data['groupsofmetricprofiles']:
-                userprofile.groupsofmetricprofiles.add(poem_models.GroupOfMetricProfiles.objects.get(name=group))
-
-            return Response(status=status.HTTP_201_CREATED)
-
-        except Exception:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
-
-    def delete(self, request, username=None):
-        if username:
-            try:
-                user = CustUser.objects.get(username=username)
-                user.delete()
-                return Response(status=status.HTTP_204_NO_CONTENT)
-
-            except CustUser.DoesNotExist:
-                raise(NotFound(status=404, detail='User not found'))
-
-        else:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
-
-
 class ListGroupsForUser(APIView):
     authentication_classes = (SessionAuthentication,)
 
@@ -413,76 +331,6 @@ class Saml2Login(APIView):
         cache.delete_many(self._prefix(self.keys))
 
         return Response(status=status.HTTP_204_NO_CONTENT)
-
-
-class GetUserprofileForUsername(APIView):
-    authentication_classes = (SessionAuthentication,)
-
-    def get(self, request, username):
-        try:
-            user = CustUser.objects.get(username=username)
-        except CustUser.DoesNotExist:
-            raise NotFound(status=404, detail='User not found')
-        else:
-            try:
-                user_profile = poem_models.UserProfile.objects.get(user=user)
-                serializer = serializers.UserProfileSerializer(user_profile)
-                return Response(serializer.data)
-
-            except poem_models.UserProfile.DoesNotExist:
-                raise NotFound(status=404, detail='User profile not found')
-
-    def put(self, request):
-        user = CustUser.objects.get(username=request.data['username'])
-        userprofile = poem_models.UserProfile.objects.get(user=user)
-        userprofile.displayname = request.data['displayname']
-        userprofile.subject = request.data['subject']
-        userprofile.egiid = request.data['egiid']
-        userprofile.save()
-
-        for group in request.data['groupsofaggregations']:
-            userprofile.groupsofaggregations.add(poem_models.GroupOfAggregations.objects.get(name=group))
-
-        for group in request.data['groupsofmetrics']:
-            userprofile.groupsofmetrics.add(poem_models.GroupOfMetrics.objects.get(name=group))
-
-        for group in request.data['groupsofmetricprofiles']:
-            userprofile.groupsofmetricprofiles.add(poem_models.GroupOfMetricProfiles.objects.get(name=group))
-
-        # remove the groups that existed before, and now were removed:
-        for group in userprofile.groupsofaggregations.all():
-            if group.name not in request.data['groupsofaggregations']:
-                userprofile.groupsofaggregations.remove(poem_models.GroupOfAggregations.objects.get(name=group))
-
-        for group in userprofile.groupsofmetrics.all():
-            if group.name not in request.data['groupsofmetrics']:
-                userprofile.groupsofmetrics.remove(poem_models.GroupOfMetrics.objects.get(name=group))
-
-        for group in userprofile.groupsofmetricprofiles.all():
-            if group.name not in request.data['groupsofmetricprofiles']:
-                userprofile.groupsofmetricprofiles.remove(poem_models.GroupOfMetricProfiles.objects.get(name=group))
-
-        return Response(status.HTTP_201_CREATED)
-
-
-class ListGroupsForGivenUser(APIView):
-    authentication_classes = (SessionAuthentication,)
-
-    def get(self, request, username=None):
-        if username:
-            try:
-                user = CustUser.objects.get(username=username)
-
-            except CustUser.DoesNotExist:
-                raise NotFound(status=404, detail='User not found')
-
-            else:
-                results = get_groups_for_user(user)
-
-        else:
-            results = get_all_groups()
-
-        return Response({'result': results})
 
 
 class ListAggregationsInGroup(APIView):
